requirements.txt
Flask-SQLAlchemy==1.0
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
virtualenv==12.0.7
Werkzeug==0.10.1
Flask-WTF==0.9.5